creating new audio recording file with every new recording
fixed thread sleep time issue for all uses of the sleep function, this shoudl result in much lower CPU load





git-svn-id: https://elphel.svn.sourceforge.net/svnroot/elphel/Apertus/ElphelVision/Java App@273 46015a06-04f4-4860-9c33-90e0f15d9639

+++ src/DatarateMonitor.java
@@ -57,7 +57,7 @@ public class DatarateMonitor extends JPanel implements Runnable, java.io.Seriali
                 repaint();
 
                 try {
-                    Thread.sleep(1 / fps * 1000);
+                    Thread.sleep((int)(1.0f / fps * 1000.0f));
                 } catch (InterruptedException e) {
                     break;
                 }
