+++ src/Histogram.java
@@ -60,7 +60,7 @@ public class Histogram extends JPanel implements Runnable, java.io.Serializable 
             repaint();
 
             try {
-                Thread.sleep(1 / fps * 1000);
+                Thread.sleep((int) (1.0f / fps * 1000.0f));
             } catch (InterruptedException e) {
                 break;
             }
